program1.java
This Java program demonstrates how to calculate the sum of the first N natural numbers using a loop. The program contains two classes: Logic and Program1. In the Logic class, the method CalculateSum(int iNo) accepts a number as input. Inside this method, two variables iCnt and iCount are declared, where iCnt is used as a loop counter and iCount stores the cumulative total. A for loop runs from 1 to the value of iNo, and during each iteration, the current number iCnt is added to iCount. After the loop completes, the method prints the final result as "sum of N Number is:". The Program1 class contains the main method, where an object of the Logic class is created, and the CalculateSum method is called with the value 10. This results in the program calculating and printing the sum of numbers from 1 to 10, which is 55. The program shows the use of looping, method calling, and class-object interaction in Java.

Signed-off-by: Apurva Satish Ratnakar <apurvasatishratnakar1409@gmail.com>