program3.java
This Java program calculates the factorial of a given number. It consists of two classes: Logic and Program3. The Logic class contains a method named FindFactorial(int iNo) that takes an integer as input. Inside the method, two variables are declared: iFact initialized to 1 for storing the factorial result, and iCnt used as a loop counter. A for loop runs from 1 to the given number iNo, and in each iteration, the current value of iCnt is multiplied with iFact. This process continues until the loop completes, resulting in iFact storing the factorial value. After the loop ends, the program prints the result. In the Program3 class, the main method creates an object of the Logic class and calls FindFactorial(5), which calculates and displays the factorial of 5. This program demonstrates the use of loops, method calls, and basic logic building in Java.

Signed-off-by: Apurva Satish Ratnakar <apurvasatishratnakar1409@gmail.com>